requirements.txt
asgiref
certifi
crispy-bootstrap5
distlib
dj-database-url
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
filelock
Markdown
MouseInfo
numpy
opencv-contrib-python
pillow
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
setuptools
sqlparse
typing_extensions
tzdata
virtualenv